Lens.Family

Description

This is the main module for end-users of lens-families-core. If you are not building your own optics such as lenses, traversals, grates, etc., but just using optics made by others, this is the only module you need.

Lenses

This module provides ^. for accessing fields and .~ and %~ for setting and modifying fields. Lenses are composed with . from the Prelude and id is the identity lens.

Lens composition in this library enjoys the following identities.

x^.l1.l2 === x^.l1^.l2

l1.l2 %~ f === l1 %~ l2 %~ f

The identity lens behaves as follows.

x^.id === x

id %~ f === f

The & operator, allows for a convenient way to sequence record updating:

record & l1 .~ value1 & l2 .~ value2

Lenses are implemented in van Laarhoven style. Lenses have type Functor f => (a -> f a) -> s -> f s and lens families have type Functor f => (a i -> f (a j)) -> s i -> f (s j).

Keep in mind that lenses and lens families can be used directly for functorial updates. For example, _2 id gives you strength.

_2 id :: Functor f => (a, f b) -> f (a, b)

Here is an example of code that uses the Maybe functor to preserves sharing during update when possible.

-- | 'sharedUpdate' returns the *identical* object if the update doesn't change anything.
-- This is useful for preserving sharing.
sharedUpdate :: Eq a => LensLike' Maybe s a -> (a -> a) -> s -> s
sharedUpdate l f s = fromMaybe s (l f' s)
 where
  f' a | b == a    = Nothing
       | otherwise = Just b
   where
    b = f a

Traversals

^. can be used with traversals to access monoidal fields. The result will be a mconcat of all the fields referenced. The various fooOf functions can be used to access different monoidal summaries of some kinds of values.

^? can be used to access the first value of a traversal. Nothing is returned when the traversal has no references.

^.. can be used with a traversals and will return a list of all fields referenced.

When .~ is used with a traversal, all referenced fields will be set to the same value, and when %~ is used with a traversal, all referenced fields will be modified with the same function.

A variant of ^? call matching returns Either a Right value which is the first value of the traversal, or a Left value which is a "proof" that the traversal has no elements. The "proof" consists of the original input structure, but in the case of polymorphic families, the type parameter is replaced with a fresh type variable, thus proving that the type parameter was unused.

Like all optics, traversals can be composed with ., and because every lens is automatically a traversal, lenses and traversals can be composed with . yielding a traversal.

Traversals are implemented in van Laarhoven style. Traversals have type Applicative f => (a -> f a) -> s -> f s and traversal families have type Applicative f => (a i -> f (a j)) -> s i -> f (s j).

Grates

zipWithOf can be used with grates to zip two structure together provided a binary operation.

under can be to modify each value in a structure according to a function. This works analogous to how over works for lenses and traversals.

review can be used with grates to construct a constant grate from a single value. This is like a 0-ary zipWith function.

degrating can be used to build higher arity zipWithOf functions:

zipWith3Of :: AGrate s t a b -> (a -> a -> a -> b) -> s -> s -> s -> t
zipWith3Of l f s1 s2 s3 = degrating l (\k -> f (k s1) (k s2) (k s3))

Like all optics, grates can be composed with ., and id is the identity grate.

Grates are implemented in van Laarhoven style.

Grates have type Functor g => (g a -> a) -> g s -> s and grate families have type Functor g => (g (a i) -> a j) -> g (s i) -> s j.

Keep in mind that grates and grate families can be used directly for functorial zipping. For example,

both sum :: Num a => [(a, a)] -> (a, a)

will take a list of pairs return the sum of the first components and the sum of the second components. For another example,

cod id :: Functor f => f (r -> a) -> r -> f a

will turn a functor full of functions into a function returning a functor full of results.

Adapters, Grids, and Prisms

The Adapter, Prism, and Grid optics are all AdapterLike optics and typically not used directly, but either converted to a LensLike optic using under, or into a GrateLike optic using over. See under and over for details about which conversions are possible.

These optics are implemented in van Laarhoven style.

Adapters have type (Functor f, Functor g) => (g a -> f a) -> g s -> f s and Adapters families have type (Functor f, Functor g) => (g (a i) -> f (a j)) -> g (s i) -> f (s j).

Grids have type (Applicative f, Functor g) => (g a -> f a) -> g s -> f s and Grids families have type (Applicative f, Functor g) => (g (a i) -> f (a j)) -> g (s i) -> f (s j).

Prisms have type (Applicative f, Traversable g) => (g a -> f a) -> g s -> f s and Prisms families have type (Applicative f, Traversable g) => (g (a i) -> f (a j)) -> g (s i) -> f (s j).

Keep in mind that these optics and their families can sometimes be used directly, without using over and under. Sometimes you can take advantage of the fact that

   LensLike f (g s) t (g a) b
  ==
   AdapterLike f g s t a b
  ==
   GrateLike g s (f t) a (f b)

For example, if you have a grid for your structure to another type that has an Arbitray instance, such as grid from a custom word type to Bool, e.g. myWordBitVector :: (Applicative f, Functor g) => AdapterLike' f g MyWord Bool, you can use the grid to create an Arbitrary instance for your structure by directly applying review:

instance Arbitrary MyWord where
  arbitrary = review myWordBitVector arbitrary

Building and Finding Optics

To build your own optics, see Lens.Family.Unchecked.

For stock optics, see Lens.Family.Stock.

References:

http://www.twanvl.nl/blog/haskell/cps-functional-references

http://r6.ca/blog/20120623T104901Z.html

http://comonad.com/reader/2012/mirrored-lenses/

http://conal.net/blog/posts/semantic-editor-combinators

https://r6research.livejournal.com/28050.html

Documentation

to :: Phantom f => (s -> a) -> LensLike f s t a b

to :: (s -> a) -> Getter s t a b

to promotes a projection function to a read-only lens called a getter. To demote a lens to a projection function, use the section (^.l) or view l.

>>> (3 :+ 4, "example")^._1.to(abs)
5.0 :+ 0.0

view :: FoldLike a s t a b -> s -> a

view :: Getter s t a b -> s -> a

Demote a lens or getter to a projection function.

view :: Monoid a => Fold s t a b -> s -> a

Returns the monoidal summary of a traversal or a fold.

(^.) :: s -> FoldLike a s t a b -> a infixl 8

(^.) :: s -> Getter s t a b -> a

Access the value referenced by a getter or lens.

(^.) :: Monoid a => s -> Fold s t a b -> a

Access the monoidal summary referenced by a traversal or a fold.

folding :: (Foldable g, Phantom f, Applicative f) => (s -> g a) -> LensLike f s t a b

folding :: (s -> [a]) -> Fold s t a b

folding promotes a "toList" function to a read-only traversal called a fold.

To demote a traversal or fold to a "toList" function use the section (^..l) or toListOf l.

views :: FoldLike r s t a b -> (a -> r) -> s -> r

views :: Monoid r => Fold s t a b -> (a -> r) -> s -> r

Given a fold or traversal, return the foldMap of all the values using the given function.

views :: Getter s t a b -> (a -> r) -> s -> r

views is not particularly useful for getters or lenses, but given a getter or lens, it returns the referenced value passed through the given function.

views l f s = f (view l s)

(^..) :: s -> FoldLike [a] s t a b -> [a] infixl 8

(^..) :: s -> Fold s t a b -> [a]

Returns a list of all of the referenced values in order.

(^?) :: s -> FoldLike (First a) s t a b -> Maybe a infixl 8

(^?) :: s -> Fold s t a b -> Maybe a

Returns Just the first referenced value. Returns Nothing if there are no referenced values.

toListOf :: FoldLike [a] s t a b -> s -> [a]

toListOf :: Fold s t a b -> s -> [a]

Returns a list of all of the referenced values in order.

allOf :: FoldLike All s t a b -> (a -> Bool) -> s -> Bool

allOf :: Fold s t a b -> (a -> Bool) -> s -> Bool

Returns true if all of the referenced values satisfy the given predicate.

anyOf :: FoldLike Any s t a b -> (a -> Bool) -> s -> Bool

anyOf :: Fold s t a b -> (a -> Bool) -> s -> Bool

Returns true if any of the referenced values satisfy the given predicate.

firstOf :: FoldLike (First a) s t a b -> s -> Maybe a

firstOf :: Fold s t a b -> s -> Maybe a

Returns Just the first referenced value. Returns Nothing if there are no referenced values. See ^? for an infix version of firstOf

lastOf :: FoldLike (Last a) s t a b -> s -> Maybe a

lastOf :: Fold s t a b -> s -> Maybe a

Returns Just the last referenced value. Returns Nothing if there are no referenced values.

sumOf :: Num a => FoldLike (Sum a) s t a b -> s -> a

sumOf :: Num a => Fold s t a b -> s -> a

Returns the sum of all the referenced values.

productOf :: Num a => FoldLike (Product a) s t a b -> s -> a

productOf :: Num a => Fold s t a b -> s -> a

Returns the product of all the referenced values.

lengthOf :: Num r => FoldLike (Sum r) s t a b -> s -> r

lengthOf :: Num r => Fold s t a b -> s -> r

Counts the number of references in a traversal or fold for the input.

nullOf :: FoldLike All s t a b -> s -> Bool

nullOf :: Fold s t a b -> s -> Bool

Returns true if the number of references in the input is zero.

matching :: LensLike (Either a) s t a b -> s -> Either t a

matching :: Traversal s t a b -> s -> Either t a

Returns Right of the first referenced value. Returns Left the original value when there are no referenced values. In case there are no referenced values, the result might have a fresh type parameter, thereby proving the original value had no referenced values.

over :: ASetter s t a b -> (a -> b) -> s -> t

over :: Setter s t a b -> (a -> b) -> s -> t

Demote a setter to a semantic editor combinator.

over :: Prism s t a b -> Reviwer s t a b
over :: Grid s t a b -> Grate s t a b
over :: Adapter s t a b -> Grate s t a b

Covert an AdapterLike optic into a GrateLike optic.

(%~) :: ASetter s t a b -> (a -> b) -> s -> t infixr 4

Modify all referenced fields.

set :: ASetter s t a b -> b -> s -> t

Set all referenced fields to the given value.

(.~) :: ASetter s t a b -> b -> s -> t infixr 4

Set all referenced fields to the given value.

review :: GrateLike (Constant ()) s t a b -> b -> t

review :: Grate s t a b -> b -> t
review :: Reviewer s t a b -> b -> t

zipWithOf :: GrateLike (Prod Identity Identity) s t a b -> (a -> a -> b) -> s -> s -> t

zipWithOf :: Grate s t a b -> (a -> a -> b) -> s -> s -> t

Returns a binary instance of a grate.

zipWithOf l f x y = degrating l (k -> f (k x) (k y))

degrating :: AGrate s t a b -> ((s -> a) -> b) -> t

degrating :: Grate s t a b -> ((s -> a) -> b) -> t

Demote a grate to its normal, higher-order function, form.

degrating . grate = id
grate . degrating = id

under :: AResetter s t a b -> (a -> b) -> s -> t

under :: Resetter s t a b -> (a -> b) -> s -> t

Demote a resetter to a semantic editor combinator.

under :: Prism s t a b -> Traversal s t a b
under :: Grid s t a b -> Traversal s t a b
under :: Adapter s t a b -> Lens s t a b

Covert an AdapterLike optic into a LensLike optic.

Note: this function is unrelated to the lens package's under function.

reset :: AResetter s t a b -> b -> s -> t

reset :: Resetter s t a b -> b -> s -> t

Set all referenced fields to the given value.

(&) :: s -> (s -> t) -> t infixl 1

A flipped version of ($).

Pseudo-imperatives

(+~) :: Num a => ASetter s t a a -> a -> s -> t infixr 4

(*~) :: Num a => ASetter s t a a -> a -> s -> t infixr 4

(-~) :: Num a => ASetter s t a a -> a -> s -> t infixr 4

(//~) :: Fractional a => ASetter s t a a -> a -> s -> t infixr 4

(&&~) :: ASetter s t Bool Bool -> Bool -> s -> t infixr 4

(||~) :: ASetter s t Bool Bool -> Bool -> s -> t infixr 4

(<>~) :: Monoid a => ASetter s t a a -> a -> s -> t infixr 4

Monoidally append a value to all referenced fields.

Types

type AdapterLike f g s t a b = (g a -> f b) -> g s -> f t

type AdapterLike' f g s a = (g a -> f a) -> g s -> f s

type LensLike f s t a b = (a -> f b) -> s -> f t

type LensLike' f s a = (a -> f a) -> s -> f s

type FoldLike r s t a b = LensLike (Constant r) s t a b

type FoldLike' r s a = LensLike' (Constant r) s a

type GrateLike g s t a b = (g a -> b) -> g s -> t

type GrateLike' g s a = (g a -> a) -> g s -> s

type AGrate s t a b = GrateLike (PCont b a) s t a b

type AGrate' s a = GrateLike' (PCont a a) s a

type ASetter s t a b = LensLike Identity s t a b

type ASetter' s a = LensLike' Identity s a

type AResetter s t a b = GrateLike Identity s t a b

type AResetter' s a = GrateLike' Identity s a

data PCont i j a

data First a

data Last a

class Functor f => Phantom f

Minimal complete definition

coerce

Re-exports

data Constant a (b :: k)

Constant functor.

data Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

type Prod = Product

data All

Boolean monoid under conjunction (&&).

>>> getAll (All True <> mempty <> All False)
False

>>> getAll (mconcat (map (\x -> All (even x)) [2,4,6,7,8]))
False

data Any

Boolean monoid under disjunction (||).

>>> getAny (Any True <> mempty <> Any False)
True

>>> getAny (mconcat (map (\x -> Any (even x)) [2,4,6,7,8]))
True

data Sum a

Monoid under addition.

>>> getSum (Sum 1 <> Sum 2 <> mempty)
3

data Product a

Monoid under multiplication.

>>> getProduct (Product 3 <> Product 4 <> mempty)
12